Improve `DEFINE` and `REMOVE` statements
