+++ lib/src/kvs/tx.rs
@@ -15,7 +15,6 @@ use crate::sql;
 use crate::sql::paths::EDGE;
 use crate::sql::paths::IN;
 use crate::sql::paths::OUT;
-use crate::sql::statements::DefineUserStatement;
 use crate::sql::thing::Thing;
 use crate::sql::Strand;
 use crate::sql::Value;
@@ -30,12 +29,12 @@ use sql::statements::DefineEventStatement;
 use sql::statements::DefineFieldStatement;
 use sql::statements::DefineFunctionStatement;
 use sql::statements::DefineIndexStatement;
-use sql::statements::DefineLoginStatement;
 use sql::statements::DefineNamespaceStatement;
 use sql::statements::DefineParamStatement;
 use sql::statements::DefineScopeStatement;
 use sql::statements::DefineTableStatement;
 use sql::statements::DefineTokenStatement;
+use sql::statements::DefineUserStatement;
 use sql::statements::LiveStatement;
 use std::borrow::Cow;
 use std::collections::HashMap;
@@ -1205,32 +1204,23 @@ impl Transaction {
 		})
 	}
 
-	/// Retrieve all namespace login definitions for a specific namespace.
-	pub async fn all_nl(&mut self, ns: &str) -> Result<Arc<[DefineLoginStatement]>, Error> {
-		let key = crate::key::namespace::lg::prefix(ns);
+	/// Retrieve all namespace user definitions for a specific namespace.
+	pub async fn all_ns_users(&mut self, ns: &str) -> Result<Arc<[DefineUserStatement]>, Error> {
+		let key = crate::key::namespace::us::prefix(ns);
 		Ok(if let Some(e) = self.cache.get(&key) {
-			if let Entry::Nls(v) = e {
+			if let Entry::Nus(v) = e {
 				v
 			} else {
 				unreachable!();
 			}
 		} else {
-			let beg = crate::key::namespace::lg::prefix(ns);
-			let end = crate::key::namespace::lg::suffix(ns);
+			let beg = crate::key::namespace::us::prefix(ns);
+			let end = crate::key::namespace::us::suffix(ns);
 			let val = self.getr(beg..end, u32::MAX).await?;
 			let val = val.convert().into();
-			self.cache.set(key, Entry::Nls(Arc::clone(&val)));
+			self.cache.set(key, Entry::Nus(Arc::clone(&val)));
 			val
 		})
-	}
-
-	/// Retrieve all namespace user definitions for a specific namespace.
-	pub async fn all_ns_users(&mut self, ns: &str) -> Result<Arc<[DefineUserStatement]>, Error> {
-		let beg = crate::key::namespace::us::prefix(ns);
-		let end = crate::key::namespace::us::suffix(ns);
-		let val = self.getr(beg..end, u32::MAX).await?;
-		let val = val.convert().into();
-		Ok(val)
 	}
 
 	/// Retrieve all namespace token definitions for a specific namespace.
@@ -1271,40 +1261,27 @@ impl Transaction {
 		})
 	}
 
-	/// Retrieve all database login definitions for a specific database.
-	pub async fn all_dl(
-		&mut self,
-		ns: &str,
-		db: &str,
-	) -> Result<Arc<[DefineLoginStatement]>, Error> {
-		let key = crate::key::database::lg::prefix(ns, db);
-		Ok(if let Some(e) = self.cache.get(&key) {
-			if let Entry::Dls(v) = e {
-				v
-			} else {
-				unreachable!();
-			}
-		} else {
-			let beg = crate::key::database::lg::prefix(ns, db);
-			let end = crate::key::database::lg::suffix(ns, db);
-			let val = self.getr(beg..end, u32::MAX).await?;
-			let val = val.convert().into();
-			self.cache.set(key, Entry::Dls(Arc::clone(&val)));
-			val
-		})
-	}
-
 	/// Retrieve all database user definitions for a specific database.
 	pub async fn all_db_users(
 		&mut self,
 		ns: &str,
 		db: &str,
 	) -> Result<Arc<[DefineUserStatement]>, Error> {
-		let beg = crate::key::database::us::prefix(ns, db);
-		let end = crate::key::database::us::suffix(ns, db);
-		let val = self.getr(beg..end, u32::MAX).await?;
-		let val = val.convert().into();
-		Ok(val)
+		let key = crate::key::database::us::prefix(ns, db);
+		Ok(if let Some(e) = self.cache.get(&key) {
+			if let Entry::Dus(v) = e {
+				v
+			} else {
+				unreachable!();
+			}
+		} else {
+			let beg = crate::key::database::us::prefix(ns, db);
+			let end = crate::key::database::us::suffix(ns, db);
+			let val = self.getr(beg..end, u32::MAX).await?;
+			let val = val.convert().into();
+			self.cache.set(key, Entry::Dus(Arc::clone(&val)));
+			val
+		})
 	}
 
 	/// Retrieve all database token definitions for a specific database.
@@ -1631,15 +1608,6 @@ impl Transaction {
 		Ok(val.into())
 	}
 
-	/// Retrieve a specific namespace login definition.
-	pub async fn get_nl(&mut self, ns: &str, nl: &str) -> Result<DefineLoginStatement, Error> {
-		let key = crate::key::namespace::lg::new(ns, nl);
-		let val = self.get(key).await?.ok_or(Error::NlNotFound {
-			value: nl.to_owned(),
-		})?;
-		Ok(val.into())
-	}
-
 	/// Retrieve a specific namespace token definition.
 	pub async fn get_nt(&mut self, ns: &str, nt: &str) -> Result<DefineTokenStatement, Error> {
 		let key = crate::key::namespace::tk::new(ns, nt);
@@ -1654,20 +1622,6 @@ impl Transaction {
 		let key = crate::key::namespace::db::new(ns, db);
 		let val = self.get(key).await?.ok_or(Error::DbNotFound {
 			value: db.to_owned(),
-		})?;
-		Ok(val.into())
-	}
-
-	/// Retrieve a specific database login definition.
-	pub async fn get_dl(
-		&mut self,
-		ns: &str,
-		db: &str,
-		dl: &str,
-	) -> Result<DefineLoginStatement, Error> {
-		let key = crate::key::database::lg::new(ns, db, dl);
-		let val = self.get(key).await?.ok_or(Error::DlNotFound {
-			value: dl.to_owned(),
 		})?;
 		Ok(val.into())
 	}
@@ -1868,7 +1822,7 @@ impl Transaction {
 					let key = crate::key::root::ns::new(ns);
 					let val = DefineNamespaceStatement {
 						name: ns.to_owned().into(),
-						id: None,
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(val)
@@ -1897,8 +1851,7 @@ impl Transaction {
 					let key = crate::key::namespace::db::new(ns, db);
 					let val = DefineDatabaseStatement {
 						name: db.to_owned().into(),
-						changefeed: None,
-						id: None,
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(val)
@@ -1928,7 +1881,7 @@ impl Transaction {
 					let key = crate::key::database::sc::new(ns, db, sc);
 					let val = DefineScopeStatement {
 						name: sc.to_owned().into(),
-						..DefineScopeStatement::default()
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(val)
@@ -1959,8 +1912,7 @@ impl Transaction {
 					let val = DefineTableStatement {
 						name: tb.to_owned().into(),
 						permissions: Permissions::none(),
-						changefeed: None,
-						..DefineTableStatement::default()
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(val)
@@ -2057,7 +2009,7 @@ impl Transaction {
 					let key = crate::key::root::ns::new(ns);
 					let val = DefineNamespaceStatement {
 						name: ns.to_owned().into(),
-						id: None,
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(Arc::new(val))
@@ -2086,8 +2038,7 @@ impl Transaction {
 					let key = crate::key::namespace::db::new(ns, db);
 					let val = DefineDatabaseStatement {
 						name: db.to_owned().into(),
-						changefeed: None,
-						id: None,
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(Arc::new(val))
@@ -2118,8 +2069,7 @@ impl Transaction {
 					let val = DefineTableStatement {
 						name: tb.to_owned().into(),
 						permissions: Permissions::none(),
-						changefeed: None,
-						..DefineTableStatement::default()
+						..Default::default()
 					};
 					self.put(key, &val).await?;
 					Ok(Arc::new(val))
@@ -2183,16 +2133,16 @@ impl Transaction {
 				chn.send(bytes!("")).await?;
 			}
 		}
-		// Output LOGINS
+		// Output USERS
 		{
-			let dls = self.all_dl(ns, db).await?;
-			if !dls.is_empty() {
+			let dus = self.all_db_users(ns, db).await?;
+			if !dus.is_empty() {
 				chn.send(bytes!("-- ------------------------------")).await?;
-				chn.send(bytes!("-- LOGINS")).await?;
+				chn.send(bytes!("-- USERS")).await?;
 				chn.send(bytes!("-- ------------------------------")).await?;
 				chn.send(bytes!("")).await?;
-				for dl in dls.iter() {
-					chn.send(bytes!(format!("{dl};"))).await?;
+				for us in dus.iter() {
+					chn.send(bytes!(format!("{us};"))).await?;
 				}
 				chn.send(bytes!("")).await?;
 			}
@@ -2383,6 +2333,13 @@ impl Transaction {
 		}
 		// Everything exported
 		Ok(())
+	}
+
+	// change will record the change in the changefeed if enabled.
+	// To actually persist the record changes into the underlying kvs,
+	// you must call the `complete_changes` function and then commit the transaction.
+	pub(crate) fn clear_cache(&mut self) {
+		self.cache.clear()
 	}
 
 	// change will record the change in the changefeed if enabled.
@@ -2800,6 +2757,9 @@ mod tests {
 		let key2 = crate::key::namespace::us::new("ns", "user2");
 		let _ = txn.set(key1, data.to_owned()).await.unwrap();
 		let _ = txn.set(key2, data.to_owned()).await.unwrap();
+
+		txn.cache.clear();
+
 		let res = txn.all_ns_users("ns").await.unwrap();
 
 		assert_eq!(res.len(), 2);
@@ -2808,7 +2768,7 @@ mod tests {
 	}
 
 	#[tokio::test]
-	async fn test_db_users() {
+	async fn test_all_db_users() {
 		let ds = Datastore::new("memory").await.unwrap();
 		let mut txn = ds.transaction(true, false).await.unwrap();
 
@@ -2827,6 +2787,9 @@ mod tests {
 		let key2 = crate::key::database::us::new("ns", "db", "user2");
 		let _ = txn.set(key1, data.to_owned()).await.unwrap();
 		let _ = txn.set(key2, data.to_owned()).await.unwrap();
+
+		txn.cache.clear();
+
 		let res = txn.all_db_users("ns", "db").await.unwrap();
 
 		assert_eq!(res.len(), 2);
