+++ lib/src/sql/mod.rs
@@ -107,6 +107,7 @@ pub use self::id::Id;
 pub use self::ident::Ident;
 pub use self::idiom::Idiom;
 pub use self::idiom::Idioms;
+pub use self::index::Index;
 pub use self::kind::Kind;
 pub use self::limit::Limit;
 pub use self::model::Model;
@@ -124,6 +125,7 @@ pub use self::permission::Permissions;
 pub use self::query::Query;
 pub use self::range::Range;
 pub use self::regex::Regex;
+pub use self::scoring::Scoring;
 pub use self::script::Script;
 pub use self::split::Split;
 pub use self::split::Splits;
