+++ lib/src/sql/statements/mod.rs
@@ -53,7 +53,6 @@ pub use self::define::DefineEventStatement;
 pub use self::define::DefineFieldStatement;
 pub use self::define::DefineFunctionStatement;
 pub use self::define::DefineIndexStatement;
-pub use self::define::DefineLoginStatement;
 pub use self::define::DefineNamespaceStatement;
 pub use self::define::DefineParamStatement;
 pub use self::define::DefineScopeStatement;
@@ -67,7 +66,6 @@ pub use self::remove::RemoveEventStatement;
 pub use self::remove::RemoveFieldStatement;
 pub use self::remove::RemoveFunctionStatement;
 pub use self::remove::RemoveIndexStatement;
-pub use self::remove::RemoveLoginStatement;
 pub use self::remove::RemoveNamespaceStatement;
 pub use self::remove::RemoveParamStatement;
 pub use self::remove::RemoveScopeStatement;
