--- lib/src/sql/statements/remove.rs
@@ -1,908 +1,0 @@
-use crate::ctx::Context;
-use crate::dbs::Options;
-use crate::dbs::Transaction;
-use crate::doc::CursorDoc;
-use crate::err::Error;
-use crate::iam::{Action, ResourceKind};
-use crate::sql::base::{base, base_or_scope, Base};
-use crate::sql::comment::{mightbespace, shouldbespace};
-use crate::sql::error::IResult;
-use crate::sql::ident;
-use crate::sql::ident::{ident, Ident};
-use crate::sql::idiom;
-use crate::sql::idiom::Idiom;
-use crate::sql::value::Value;
-use derive::Store;
-use nom::branch::alt;
-use nom::bytes::complete::tag;
-use nom::bytes::complete::tag_no_case;
-use nom::character::complete::char;
-use nom::combinator::{map, opt};
-use nom::sequence::tuple;
-use revision::revisioned;
-use serde::{Deserialize, Serialize};
-use std::fmt::{self, Display, Formatter};
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub enum RemoveStatement {
-	Namespace(RemoveNamespaceStatement),
-	Database(RemoveDatabaseStatement),
-	Function(RemoveFunctionStatement),
-	Analyzer(RemoveAnalyzerStatement),
-	Login(RemoveLoginStatement),
-	Token(RemoveTokenStatement),
-	Scope(RemoveScopeStatement),
-	Param(RemoveParamStatement),
-	Table(RemoveTableStatement),
-	Event(RemoveEventStatement),
-	Field(RemoveFieldStatement),
-	Index(RemoveIndexStatement),
-	User(RemoveUserStatement),
-}
-
-impl RemoveStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-		_doc: Option<&CursorDoc<'_>>,
-	) -> Result<Value, Error> {
-		match self {
-			Self::Namespace(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Database(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Function(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Login(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Token(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Scope(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Param(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Table(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Event(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Field(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Index(ref v) => v.compute(ctx, opt, txn).await,
-			Self::Analyzer(ref v) => v.compute(ctx, opt, txn).await,
-			Self::User(ref v) => v.compute(ctx, opt, txn).await,
-		}
-	}
-}
-
-impl Display for RemoveStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		match self {
-			Self::Namespace(v) => Display::fmt(v, f),
-			Self::Database(v) => Display::fmt(v, f),
-			Self::Function(v) => Display::fmt(v, f),
-			Self::Login(v) => Display::fmt(v, f),
-			Self::Token(v) => Display::fmt(v, f),
-			Self::Scope(v) => Display::fmt(v, f),
-			Self::Param(v) => Display::fmt(v, f),
-			Self::Table(v) => Display::fmt(v, f),
-			Self::Event(v) => Display::fmt(v, f),
-			Self::Field(v) => Display::fmt(v, f),
-			Self::Index(v) => Display::fmt(v, f),
-			Self::Analyzer(v) => Display::fmt(v, f),
-			Self::User(v) => Display::fmt(v, f),
-		}
-	}
-}
-
-pub fn remove(i: &str) -> IResult<&str, RemoveStatement> {
-	alt((
-		map(namespace, RemoveStatement::Namespace),
-		map(database, RemoveStatement::Database),
-		map(function, RemoveStatement::Function),
-		map(login, RemoveStatement::Login),
-		map(token, RemoveStatement::Token),
-		map(scope, RemoveStatement::Scope),
-		map(param, RemoveStatement::Param),
-		map(table, RemoveStatement::Table),
-		map(event, RemoveStatement::Event),
-		map(field, RemoveStatement::Field),
-		map(index, RemoveStatement::Index),
-		map(analyzer, RemoveStatement::Analyzer),
-		map(user, RemoveStatement::User),
-	))(i)
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveNamespaceStatement {
-	pub name: Ident,
-}
-
-impl RemoveNamespaceStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Namespace, &Base::Root)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::root::ns::new(&self.name);
-		run.del(key).await?;
-		// Delete the resource data
-		let key = crate::key::namespace::all::new(&self.name);
-		run.delp(key, u32::MAX).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveNamespaceStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE NAMESPACE {}", self.name)
-	}
-}
-
-fn namespace(i: &str) -> IResult<&str, RemoveNamespaceStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = alt((tag_no_case("NS"), tag_no_case("NAMESPACE")))(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveNamespaceStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveDatabaseStatement {
-	pub name: Ident,
-}
-
-impl RemoveDatabaseStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Database, &Base::Ns)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::namespace::db::new(opt.ns(), &self.name);
-		run.del(key).await?;
-		// Delete the resource data
-		let key = crate::key::database::all::new(opt.ns(), &self.name);
-		run.delp(key, u32::MAX).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveDatabaseStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE DATABASE {}", self.name)
-	}
-}
-
-fn database(i: &str) -> IResult<&str, RemoveDatabaseStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = alt((tag_no_case("DB"), tag_no_case("DATABASE")))(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveDatabaseStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveFunctionStatement {
-	pub name: Ident,
-}
-
-impl RemoveFunctionStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Function, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::database::fc::new(opt.ns(), opt.db(), &self.name);
-		run.del(key).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveFunctionStatement {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		write!(f, "REMOVE FUNCTION fn::{}", self.name)
-	}
-}
-
-fn function(i: &str) -> IResult<&str, RemoveFunctionStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("FUNCTION")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag("fn::")(i)?;
-	let (i, name) = ident::plain(i)?;
-	let (i, _) = opt(|i| {
-		let (i, _) = mightbespace(i)?;
-		let (i, _) = char('(')(i)?;
-		let (i, _) = mightbespace(i)?;
-		let (i, _) = char(')')(i)?;
-		Ok((i, ()))
-	})(i)?;
-	Ok((
-		i,
-		RemoveFunctionStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveAnalyzerStatement {
-	pub name: Ident,
-}
-
-impl RemoveAnalyzerStatement {
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Analyzer, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::database::az::new(opt.ns(), opt.db(), &self.name);
-		run.del(key).await?;
-		// TODO Check that the analyzer is not used in any schema
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveAnalyzerStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE ANALYZER {}", self.name)
-	}
-}
-
-fn analyzer(i: &str) -> IResult<&str, RemoveAnalyzerStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ANALYZER")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveAnalyzerStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveLoginStatement {
-	pub name: Ident,
-	pub base: Base,
-}
-
-impl RemoveLoginStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Actor, &self.base)?;
-
-		match self.base {
-			Base::Ns => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::namespace::lg::new(opt.ns(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			Base::Db => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::database::lg::new(opt.ns(), opt.db(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			_ => Err(Error::InvalidLevel(self.base.to_string())),
-		}
-	}
-}
-
-impl Display for RemoveLoginStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE LOGIN {} ON {}", self.name, self.base)
-	}
-}
-
-fn login(i: &str) -> IResult<&str, RemoveLoginStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("LOGIN")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, base) = base(i)?;
-	Ok((
-		i,
-		RemoveLoginStatement {
-			name,
-			base,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveTokenStatement {
-	pub name: Ident,
-	pub base: Base,
-}
-
-impl RemoveTokenStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Actor, &self.base)?;
-
-		match &self.base {
-			Base::Ns => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::namespace::tk::new(opt.ns(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			Base::Db => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::database::tk::new(opt.ns(), opt.db(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			Base::Sc(sc) => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::scope::tk::new(opt.ns(), opt.db(), sc, &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			_ => Err(Error::InvalidLevel(self.base.to_string())),
-		}
-	}
-}
-
-impl Display for RemoveTokenStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE TOKEN {} ON {}", self.name, self.base)
-	}
-}
-
-fn token(i: &str) -> IResult<&str, RemoveTokenStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("TOKEN")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, base) = base_or_scope(i)?;
-	Ok((
-		i,
-		RemoveTokenStatement {
-			name,
-			base,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveScopeStatement {
-	pub name: Ident,
-}
-
-impl RemoveScopeStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Scope, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::database::sc::new(opt.ns(), opt.db(), &self.name);
-		run.del(key).await?;
-		// Remove the resource data
-		let key = crate::key::scope::all::new(opt.ns(), opt.db(), &self.name);
-		run.delp(key, u32::MAX).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveScopeStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE SCOPE {}", self.name)
-	}
-}
-
-fn scope(i: &str) -> IResult<&str, RemoveScopeStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("SCOPE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveScopeStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveParamStatement {
-	pub name: Ident,
-}
-
-impl RemoveParamStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Parameter, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::database::pa::new(opt.ns(), opt.db(), &self.name);
-		run.del(key).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveParamStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE PARAM {}", self.name)
-	}
-}
-
-fn param(i: &str) -> IResult<&str, RemoveParamStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("PARAM")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = char('$')(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveParamStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveTableStatement {
-	pub name: Ident,
-}
-
-impl RemoveTableStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Table, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::database::tb::new(opt.ns(), opt.db(), &self.name);
-		run.del(key).await?;
-		// Remove the resource data
-		let key = crate::key::table::all::new(opt.ns(), opt.db(), &self.name);
-		run.delp(key, u32::MAX).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveTableStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE TABLE {}", self.name)
-	}
-}
-
-fn table(i: &str) -> IResult<&str, RemoveTableStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("TABLE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	Ok((
-		i,
-		RemoveTableStatement {
-			name,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveEventStatement {
-	pub name: Ident,
-	pub what: Ident,
-}
-
-impl RemoveEventStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Event, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::table::ev::new(opt.ns(), opt.db(), &self.what, &self.name);
-		run.del(key).await?;
-		// Clear the cache
-		let key = crate::key::table::ev::prefix(opt.ns(), opt.db(), &self.what);
-		run.clr(key).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveEventStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE EVENT {} ON {}", self.name, self.what)
-	}
-}
-
-fn event(i: &str) -> IResult<&str, RemoveEventStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("EVENT")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = opt(tuple((shouldbespace, tag_no_case("TABLE"))))(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, what) = ident(i)?;
-	Ok((
-		i,
-		RemoveEventStatement {
-			name,
-			what,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveFieldStatement {
-	pub name: Idiom,
-	pub what: Ident,
-}
-
-impl RemoveFieldStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Field, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let fd = self.name.to_string();
-		let key = crate::key::table::fd::new(opt.ns(), opt.db(), &self.what, &fd);
-		run.del(key).await?;
-		// Clear the cache
-		let key = crate::key::table::fd::prefix(opt.ns(), opt.db(), &self.what);
-		run.clr(key).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveFieldStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE FIELD {} ON {}", self.name, self.what)
-	}
-}
-
-fn field(i: &str) -> IResult<&str, RemoveFieldStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("FIELD")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = idiom::local(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = opt(tuple((shouldbespace, tag_no_case("TABLE"))))(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, what) = ident(i)?;
-	Ok((
-		i,
-		RemoveFieldStatement {
-			name,
-			what,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveIndexStatement {
-	pub name: Ident,
-	pub what: Ident,
-}
-
-impl RemoveIndexStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Index, &Base::Db)?;
-		// Claim transaction
-		let mut run = txn.lock().await;
-		// Delete the definition
-		let key = crate::key::table::ix::new(opt.ns(), opt.db(), &self.what, &self.name);
-		run.del(key).await?;
-		// Remove the index data
-		let key = crate::key::index::all::new(opt.ns(), opt.db(), &self.what, &self.name);
-		run.delp(key, u32::MAX).await?;
-		// Clear the cache
-		let key = crate::key::table::ix::prefix(opt.ns(), opt.db(), &self.what);
-		run.clr(key).await?;
-		// Ok all good
-		Ok(Value::None)
-	}
-}
-
-impl Display for RemoveIndexStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE INDEX {} ON {}", self.name, self.what)
-	}
-}
-
-fn index(i: &str) -> IResult<&str, RemoveIndexStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("INDEX")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = opt(tuple((shouldbespace, tag_no_case("TABLE"))))(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, what) = ident(i)?;
-	Ok((
-		i,
-		RemoveIndexStatement {
-			name,
-			what,
-		},
-	))
-}
-
-// --------------------------------------------------
-// --------------------------------------------------
-// --------------------------------------------------
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
-#[revisioned(revision = 1)]
-pub struct RemoveUserStatement {
-	pub name: Ident,
-	pub base: Base,
-}
-
-impl RemoveUserStatement {
-	/// Process this type returning a computed simple Value
-	pub(crate) async fn compute(
-		&self,
-		_ctx: &Context<'_>,
-		opt: &Options,
-		txn: &Transaction,
-	) -> Result<Value, Error> {
-		// Allowed to run?
-		opt.is_allowed(Action::Edit, ResourceKind::Actor, &self.base)?;
-
-		match self.base {
-			Base::Root => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Process the statement
-				let key = crate::key::root::us::new(&self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			Base::Ns => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::namespace::us::new(opt.ns(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			Base::Db => {
-				// Claim transaction
-				let mut run = txn.lock().await;
-				// Delete the definition
-				let key = crate::key::database::us::new(opt.ns(), opt.db(), &self.name);
-				run.del(key).await?;
-				// Ok all good
-				Ok(Value::None)
-			}
-			_ => Err(Error::InvalidLevel(self.base.to_string())),
-		}
-	}
-}
-
-impl Display for RemoveUserStatement {
-	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "REMOVE USER {} ON {}", self.name, self.base)
-	}
-}
-
-fn user(i: &str) -> IResult<&str, RemoveUserStatement> {
-	let (i, _) = tag_no_case("REMOVE")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("USER")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, name) = ident(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, _) = tag_no_case("ON")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, base) = base(i)?;
-	Ok((
-		i,
-		RemoveUserStatement {
-			name,
-			base,
-		},
-	))
-}
-
-#[cfg(test)]
-mod tests {
-
-	use super::*;
-
-	#[test]
-	fn check_remove_serialize() {
-		let stm = RemoveStatement::Namespace(RemoveNamespaceStatement {
-			name: Ident::from("test"),
-		});
-		let enc: Vec<u8> = stm.try_into().unwrap();
-		assert_eq!(9, enc.len());
-	}
-}
