+++ lib/src/sql/statements/remove/event.rs
@@ -1,0 +1,75 @@
+use crate::ctx::Context;
+use crate::dbs::Options;
+use crate::dbs::Transaction;
+use crate::err::Error;
+use crate::iam::{Action, ResourceKind};
+use crate::sql::base::Base;
+use crate::sql::comment::shouldbespace;
+use crate::sql::error::IResult;
+use crate::sql::ident::{ident, Ident};
+use crate::sql::value::Value;
+use derive::Store;
+use nom::bytes::complete::tag_no_case;
+use nom::combinator::opt;
+use nom::sequence::tuple;
+use revision::revisioned;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display, Formatter};
+
+#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
+#[revisioned(revision = 1)]
+pub struct RemoveEventStatement {
+	pub name: Ident,
+	pub what: Ident,
+}
+
+impl RemoveEventStatement {
+	/// Process this type returning a computed simple Value
+	pub(crate) async fn compute(
+		&self,
+		_ctx: &Context<'_>,
+		opt: &Options,
+		txn: &Transaction,
+	) -> Result<Value, Error> {
+		// Allowed to run?
+		opt.is_allowed(Action::Edit, ResourceKind::Event, &Base::Db)?;
+		// Claim transaction
+		let mut run = txn.lock().await;
+		// Clear the cache
+		run.clear_cache();
+		// Delete the definition
+		let key = crate::key::table::ev::new(opt.ns(), opt.db(), &self.what, &self.name);
+		run.del(key).await?;
+		// Clear the cache
+		let key = crate::key::table::ev::prefix(opt.ns(), opt.db(), &self.what);
+		run.clr(key).await?;
+		// Ok all good
+		Ok(Value::None)
+	}
+}
+
+impl Display for RemoveEventStatement {
+	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+		write!(f, "REMOVE EVENT {} ON {}", self.name, self.what)
+	}
+}
+
+pub fn event(i: &str) -> IResult<&str, RemoveEventStatement> {
+	let (i, _) = tag_no_case("REMOVE")(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, _) = tag_no_case("EVENT")(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, name) = ident(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, _) = tag_no_case("ON")(i)?;
+	let (i, _) = opt(tuple((shouldbespace, tag_no_case("TABLE"))))(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, what) = ident(i)?;
+	Ok((
+		i,
+		RemoveEventStatement {
+			name,
+			what,
+		},
+	))
+}
