+++ lib/src/sql/statements/remove/namespace.rs
@@ -1,0 +1,67 @@
+use crate::ctx::Context;
+use crate::dbs::Options;
+use crate::dbs::Transaction;
+use crate::err::Error;
+use crate::iam::{Action, ResourceKind};
+use crate::sql::base::Base;
+use crate::sql::comment::shouldbespace;
+use crate::sql::error::IResult;
+use crate::sql::ident::{ident, Ident};
+use crate::sql::value::Value;
+use derive::Store;
+use nom::branch::alt;
+use nom::bytes::complete::tag_no_case;
+use revision::revisioned;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display, Formatter};
+
+#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Store, Hash)]
+#[revisioned(revision = 1)]
+pub struct RemoveNamespaceStatement {
+	pub name: Ident,
+}
+
+impl RemoveNamespaceStatement {
+	/// Process this type returning a computed simple Value
+	pub(crate) async fn compute(
+		&self,
+		_ctx: &Context<'_>,
+		opt: &Options,
+		txn: &Transaction,
+	) -> Result<Value, Error> {
+		// Allowed to run?
+		opt.is_allowed(Action::Edit, ResourceKind::Namespace, &Base::Root)?;
+		// Claim transaction
+		let mut run = txn.lock().await;
+		// Clear the cache
+		run.clear_cache();
+		// Delete the definition
+		let key = crate::key::root::ns::new(&self.name);
+		run.del(key).await?;
+		// Delete the resource data
+		let key = crate::key::namespace::all::new(&self.name);
+		run.delp(key, u32::MAX).await?;
+		// Ok all good
+		Ok(Value::None)
+	}
+}
+
+impl Display for RemoveNamespaceStatement {
+	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+		write!(f, "REMOVE NAMESPACE {}", self.name)
+	}
+}
+
+pub fn namespace(i: &str) -> IResult<&str, RemoveNamespaceStatement> {
+	let (i, _) = tag_no_case("REMOVE")(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, _) = alt((tag_no_case("NS"), tag_no_case("NAMESPACE")))(i)?;
+	let (i, _) = shouldbespace(i)?;
+	let (i, name) = ident(i)?;
+	Ok((
+		i,
+		RemoveNamespaceStatement {
+			name,
+		},
+	))
+}
