+++ lib/src/sql/tokenizer.rs
@@ -1,4 +1,3 @@
-use crate::sql::comment::shouldbespace;
 use crate::sql::common::commas;
 use crate::sql::error::IResult;
 use nom::branch::alt;
@@ -41,8 +40,5 @@ fn tokenizer(i: &str) -> IResult<&str, Tokenizer> {
 }
 
 pub(super) fn tokenizers(i: &str) -> IResult<&str, Vec<Tokenizer>> {
-	let (i, _) = tag_no_case("TOKENIZERS")(i)?;
-	let (i, _) = shouldbespace(i)?;
-	let (i, t) = separated_list1(commas, tokenizer)(i)?;
-	Ok((i, t))
+	separated_list1(commas, tokenizer)(i)
 }
